Project/requirements.txt
pandas==2.2.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
faker==22.4.0
matplotlib==3.8.2
scipy==1.12.0
scikit-learn==1.4.0